% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet.R
\name{as.sheet}
\alias{as.sheet}
\title{Convert data.frame to object of class 'sheet'}
\usage{
as.sheet(data, ...)
}
\arguments{
\item{data}{data.frame/vector etc.}

\item{...}{additionla parameters for future versions}
}
\value{
object of class 'sheet' (inherited from data.frame)
}
\description{
Convert data.frame to object of class 'sheet'
}
\examples{
1
}

