% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate/recalculate sheet (data.frame)}
\usage{
calculate(sheet, ...)
}
\arguments{
\item{sheet}{object of class 'sheet'}

\item{...}{additional parameters}
}
\value{
original object of class 'sheet' with recalculated formulas
}
\description{
Calculate/recalculate sheet (data.frame)
}
\examples{
1
}

